completed validation

+++ src/main/java/lk/ijse/simple_hostel_management_hibernate/controller/ReservationFormController.java
@@ -1,13 +1,6 @@
 package lk.ijse.simple_hostel_management_hibernate.controller;
 
 import com.jfoenix.controls.JFXButton;
-
-import java.io.IOException;
-import java.net.URL;
-import java.time.LocalDate;
-import java.util.List;
-import java.util.ResourceBundle;
-
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
@@ -20,16 +13,23 @@ import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
+import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 import lk.ijse.simple_hostel_management_hibernate.controller.util.AlertController;
+import lk.ijse.simple_hostel_management_hibernate.controller.util.ValidateFields;
 import lk.ijse.simple_hostel_management_hibernate.dto.ReservationDTO;
 import lk.ijse.simple_hostel_management_hibernate.dto.StudentDTO;
 import lk.ijse.simple_hostel_management_hibernate.service.ServiceFactory;
 import lk.ijse.simple_hostel_management_hibernate.service.custom.ReservationService;
 import lk.ijse.simple_hostel_management_hibernate.view.tm.ReservationTM;
-import lk.ijse.simple_hostel_management_hibernate.view.tm.RoomTM;
+
+import java.io.IOException;
+import java.net.URL;
+import java.time.LocalDate;
+import java.util.List;
+import java.util.ResourceBundle;
 
 public class ReservationFormController {
 
@@ -141,7 +141,7 @@ public class ReservationFormController {
         ImageView imageView = new ImageView(image);
         btnBack.setGraphic(imageView);
 
-        cmbPaymentStatus.getItems().addAll("PAID","NOT PAID");
+        cmbPaymentStatus.getItems().addAll("PAID", "NOT PAID");
 
         loadStIds();
         loadRoomTypeIds();
@@ -150,7 +150,7 @@ public class ReservationFormController {
         setCellValueFactory();
     }
 
-    public void loadStIds(){
+    public void loadStIds() {
         try {
             ObservableList<String> obList = FXCollections.observableArrayList();
 
@@ -166,7 +166,7 @@ public class ReservationFormController {
         }
     }
 
-    public void loadRoomTypeIds(){
+    public void loadRoomTypeIds() {
         try {
             ObservableList<String> obList = FXCollections.observableArrayList();
 
@@ -198,45 +198,66 @@ public class ReservationFormController {
         cmbStudentId.setValue(columns.get(2).getCellData(row).toString());
         cmbRoomTypeId.setValue(columns.get(3).getCellData(row).toString());
         cmbPaymentStatus.setValue(columns.get(4).getCellData(row).toString());
+
+        txtResId.setStyle("-fx-text-fill: black; -fx-background-color: #ebebeb; -fx-background-radius: 15");
+
+        btnSave.setDisable(false);
+        btnUpdate.setDisable(false);
+        btnDelete.setDisable(false);
     }
 
     public void btnDeleteOnAction(ActionEvent actionEvent) {
-        ReservationDTO reservationDTO = getDetailsInTextFields();
-        boolean success = reservationService.deleteReservation(reservationDTO);
-        if(success){
-            AlertController.confirmmessage("reservation details deleted successfully");
-            setDataToTableView();
-            clearTxtFields();
+        boolean noEmptyFields = noEmptyValuesInTextFields();
+        if (noEmptyFields) {
+            ReservationDTO reservationDTO = getDetailsInTextFields();
+            boolean success = reservationService.deleteReservation(reservationDTO);
+            if (success) {
+                AlertController.confirmmessage("reservation details deleted successfully");
+                setDataToTableView();
+                clearTxtFields();
+            } else {
+                AlertController.errormessage("reservation details deletion process unsuccessful");
+            }
         }else{
-            AlertController.errormessage("reservation details deletion process unsuccessful");
+            AlertController.errormessage("please make sure to fill out all the required fields");
         }
     }
 
     public void btnUpdateOnAction(ActionEvent actionEvent) {
-        ReservationDTO reservationDTO = getDetailsInTextFields();
-        boolean success = reservationService.updateReservation(reservationDTO);
-        if(success){
-            AlertController.confirmmessage("reservation details updated successfully");
-            setDataToTableView();
-            clearTxtFields();
+        boolean noEmptyFields = noEmptyValuesInTextFields();
+        if (noEmptyFields) {
+            ReservationDTO reservationDTO = getDetailsInTextFields();
+            boolean success = reservationService.updateReservation(reservationDTO);
+            if (success) {
+                AlertController.confirmmessage("reservation details updated successfully");
+                setDataToTableView();
+                clearTxtFields();
+            } else {
+                AlertController.errormessage("reservation details updating process unsuccessful");
+            }
         }else{
-            AlertController.errormessage("reservation details updating process unsuccessful");
+            AlertController.errormessage("please make sure to fill out all the required fields");
         }
     }
 
     public void btnSaveOnAction(ActionEvent actionEvent) {
-        ReservationDTO reservationDTO = getDetailsInTextFields();
-        boolean success = reservationService.saveReservation(reservationDTO);
-        if(success){
-            AlertController.confirmmessage("reservation details saved successfully");
-            setDataToTableView();
-            clearTxtFields();
+        boolean noEmptyFields = noEmptyValuesInTextFields();
+        if (noEmptyFields) {
+            ReservationDTO reservationDTO = getDetailsInTextFields();
+            boolean success = reservationService.saveReservation(reservationDTO);
+            if (success) {
+                AlertController.confirmmessage("reservation details saved successfully");
+                setDataToTableView();
+                clearTxtFields();
+            } else {
+                AlertController.errormessage("reservation details saving process unsuccessful");
+            }
         }else{
-            AlertController.errormessage("reservation details saving process unsuccessful");
+            AlertController.errormessage("please make sure to fill out all the required fields");
         }
     }
 
-    public ReservationDTO getDetailsInTextFields(){
+    public ReservationDTO getDetailsInTextFields() {
         ReservationDTO reservationDTO = new ReservationDTO();
         reservationDTO.setDate(dpDate.getValue());
         reservationDTO.setReservationId(txtResId.getText());
@@ -246,7 +267,7 @@ public class ReservationFormController {
         return reservationDTO;
     }
 
-    void clearTxtFields(){
+    void clearTxtFields() {
         dpDate.setValue(null);
         txtResId.setText("");
         cmbRoomTypeId.setValue("");
@@ -257,7 +278,7 @@ public class ReservationFormController {
     private void setDataToTableView() {
         ObservableList<ReservationDTO> reservationDtoList = reservationService.getDetailsToTableView();
         ObservableList<ReservationTM> reservationTmList = FXCollections.observableArrayList();
-        for (ReservationDTO dto : reservationDtoList){
+        for (ReservationDTO dto : reservationDtoList) {
             reservationTmList.add(
                     new ReservationTM(
                             dto.getReservationId(),
@@ -271,7 +292,7 @@ public class ReservationFormController {
         tableReservation.setItems(reservationTmList);
     }
 
-    public void setCellValueFactory(){
+    public void setCellValueFactory() {
         colResId.setCellValueFactory(new PropertyValueFactory<>("reservationId"));
         colDate.setCellValueFactory(new PropertyValueFactory<>("date"));
         colStId.setCellValueFactory(new PropertyValueFactory<>("studentId"));
@@ -282,7 +303,7 @@ public class ReservationFormController {
     public void cmbRoomTypeIdOnAction(ActionEvent actionEvent) {
         String roomTypeId = cmbRoomTypeId.getValue();
 
-        if(!roomTypeId.isEmpty()) {
+        if (!roomTypeId.isEmpty()) {
             List<Integer> list = reservationService.getAvailableRoomsCount(roomTypeId);
 
             if (!list.isEmpty()) {
@@ -294,34 +315,34 @@ public class ReservationFormController {
                 lblRoomAvailability.setText(availableRooms + " rooms available");
                 lblRoomAvailability.setVisible(true);
 
-                if(availableRooms!=0) {
-                    if(availableRooms==1){
+                if (availableRooms != 0) {
+                    if (availableRooms == 1) {
                         lblRoomAvailability.setText("1 room available");
                         lblRoomAvailability.setVisible(true);
                         lblRoomAvailability.setStyle("-fx-text-fill: white");
                         btnSave.setDisable(false);
-                    }else{
+                    } else {
                         lblRoomAvailability.setText(availableRooms + " rooms available");
                         lblRoomAvailability.setVisible(true);
                         lblRoomAvailability.setStyle("-fx-text-fill: white");
                         btnSave.setDisable(false);
                     }
-                }else{
+                } else {
                     lblRoomAvailability.setText("*no rooms available*");
                     lblRoomAvailability.setVisible(true);
                     lblRoomAvailability.setStyle("-fx-text-fill: red");
                     btnSave.setDisable(true);
                 }
 
-                if(perInOtherRoom!=0) {
-                    if(perInOtherRoom==1){
+                if (perInOtherRoom != 0) {
+                    if (perInOtherRoom == 1) {
                         lblExtraPersons.setText("there is a room with 1 person already in it");
                         lblExtraPersons.setVisible(true);
-                    }else{
+                    } else {
                         lblExtraPersons.setText("there is a room with " + perInOtherRoom + " persons already in it");
                         lblExtraPersons.setVisible(true);
                     }
-                }else{
+                } else {
                     lblExtraPersons.setVisible(false);
                 }
             }
@@ -376,5 +397,44 @@ public class ReservationFormController {
         }
         stage.centerOnScreen();
         stage.show();
+    }
+
+    public boolean noEmptyValuesInTextFields() {
+        String resId = txtResId.getText();
+        LocalDate resDate = dpDate.getValue();
+        String stId = cmbStudentId.getValue();
+        String roomTypeId = cmbRoomTypeId.getValue();
+        String status = cmbPaymentStatus.getValue();
+        if (!resId.isEmpty() && resDate != null && stId != null && roomTypeId != null && status != null && !stId.isEmpty()
+                && !roomTypeId.isEmpty() && !status.isEmpty()) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    public void btnEnable() {
+        if (ValidateFields.rerservationIdCheck(txtResId.getText())) {
+            btnSave.setDisable(false);
+            btnUpdate.setDisable(false);
+            btnDelete.setDisable(false);
+        }
+    }
+
+    public void txtResIdOnMouseKeytyped(KeyEvent keyEvent) {
+        String txt = txtResId.getText();
+        if (ValidateFields.rerservationIdCheck(txt)) {
+            txtResId.setStyle("-fx-text-fill: black; -fx-background-color: #ebebeb; -fx-background-radius: 15");
+            btnEnable();
+        } else {
+            txtResId.setStyle("-fx-text-fill: red; -fx-background-color: #ebebeb; -fx-background-radius: 15");
+            btnSave.setDisable(true);
+            btnUpdate.setDisable(true);
+            btnDelete.setDisable(true);
+        }
+    }
+
+    public void btnClearOnAction(ActionEvent actionEvent) {
+        clearTxtFields();
     }
 }
